step1 refactor(bowling) NormalFrame 에서 Strike추출
