+++ src/main/java/bowling/FallDownPins.java
@@ -1,11 +1,13 @@
 package bowling;
 
+import bowling.state.State;
+import bowling.state.Strike;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
-public class FallDownPins implements State{
+public class FallDownPins implements State {
 
   private static final int FIRST_FALL_DOWN_COUNT_INDEX = 0;
   private static final int SECOND_FALL_DOWN_COUNT_INDEX = 1;
@@ -14,11 +16,14 @@ public class FallDownPins implements State{
 
   List<FallDownPin> fallDownCounts = new ArrayList<>();
 
-  public FallDownPins roll(int fallDownCount) {
+  public State roll(int fallDownCount) {
     if (!validateSumOfFallDownCount(fallDownCount)) {
       throw new IllegalArgumentException("한 프레임에서 넘어뜨릴 수 있는 핀의 합은 10을 넘어 갈 수 없습니다.");
     }
     fallDownCounts.add(FallDownPin.of(fallDownCount));
+    if (isStrike()) {
+      return new Strike();
+    }
     return this;
   }
 
@@ -71,9 +76,6 @@ public class FallDownPins implements State{
 
   @Override
   public String toString() {
-    if (isStrike()) {
-      return "X";
-    }
     if (isSpare()) {
       return getFirstFallDown().toString() + "|" + "/";
     }
