+++ src/main/java/bowling/state/State.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.state;
 
 public interface State {
